fix: CookieOverflow error on new user create
The ActiveRecord returned by the `create` operation is larger than
the ActiveRecord retrieved by the `where` lookup.   This is because of
how the `User` model attributes were stored in both cases.

The end result of the larger ActiveRecord was a CookieOverflow when
Devise attempted to serialize and store the entire `User` model
in a cookie, leading to me not being able to stop thinking about this
problem for a solid 24 hours.